test(langgraph): consolidate interrupt-graph fixtures and tighten comments
Follow-up cleanup on the stream-abandonment fix, no behavior change.

- Move the interrupt-graph construction into shared conftest fixtures
  (interrupt_graph, async_interrupt_graph), modeled on simple_graph, and
  rewire the sync/async/llmobs tests to use them. Removes the standalone
  helper and the llmobs test's inline copy.
- Drop test_astream_break_without_close_emits_trace: its GeneratorExit
  finally path is already pinned deterministically by the close()/aclose()
  teardown tests, and it was the only GC-timing-dependent (flaky) test.
- Trim the finally/AIDEV comments in patch.py and drop the ticket IDs from
  test docstrings and thread_id values.
